Add "show-hide" key to show_on arg
Add the ability to set “show-hide” as key for the show_on argument.
This would take an array of page templates as the matching value pair.

The meta box would be generated for all pages. However it would be
hidden by $.hide() unless one of the declared page-templates is
selected in the “Page Attributes” Template select menu. At this point
it is displayed by $.show();